# docs/swagger.yaml
basePath: /
definitions:
...
  response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      timestamp:
        description: Useful for frontend debugging / UX
        type: string
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  supplier.CreateRequest:
    properties:
      address:
        type: string
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      phone:
        maxLength: 50
        type: string
    required:
    - name
    type: object
  supplier.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/supplier.Supplier'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  supplier.Supplier:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
  supplier.UpdateRequest:
    properties:
      address:
        type: string
      email:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      phone:
        maxLength: 50
        type: string
    type: object
...
info:
  contact: {}
  description: Market backend (products, stock, income, sales)
  title: Hezzet Market API
  version: "1.0"
paths:
...
  /suppliers:
    get:
      description: Get paginated list of active suppliers with optional search by
        name/phone/email
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Skip (legacy, default 0). If provided, overrides page/offset.
        in: query
        name: skip
        type: integer
      - description: Search by name, phone, or email
        in: query
        name: search
        type: string
      - description: Order by field (name, created_at)
        enum:
        - name
        - created_at
        in: query
        name: order_by
        type: string
      - description: Order direction (asc/desc)
        enum:
        - asc
        - desc
        in: query
        name: order_direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/supplier.ListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List suppliers
      tags:
      - Suppliers
    post:
      consumes:
      - application/json
      description: Create a new supplier
      parameters:
      - description: Supplier data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/supplier.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/supplier.Supplier'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create supplier
      tags:
      - Suppliers
  /suppliers/{id}:
    delete:
      description: Soft delete supplier by setting is_active=false
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete supplier (soft)
      tags:
      - Suppliers
    get:
      description: Get a single supplier by its ID
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/supplier.Supplier'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get supplier by ID
      tags:
      - Suppliers
    patch:
      consumes:
      - application/json
      description: Update supplier fields (name, phone, email, address, is_active)
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/supplier.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/supplier.Supplier'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update supplier
      tags:
      - Suppliers
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"